Removed more file system calls + unused imports + moved XML utils to xml package
